requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
google-api-python-client>=2.141.0
google-auth>=2.33.0
pandas>=2.2.0
openpyxl>=3.1.0
